requirements.txt
-e .
aws-cdk.aws-dynamodb